hw1/py/requirements.txt
cycler==0.10.0
decorator==4.4.2
future==0.18.2
joblib==0.15.1
kiwisolver==1.2.0
matplotlib==3.2.2
networkx==2.4
numpy==1.19.0
opt-einsum==3.2.1
pandas==1.0.5
patsy==0.5.1
pgmpy==0.1.10
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl==1.3.1
python-dateutil==2.8.1
pytz==2020.1
scipy==1.5.0
six==1.15.0
statsmodels==0.11.1
torch==1.5.1
tqdm==4.46.1
